🦜 Update Action Button

+++ src/components/action-button.tsx
@@ -1,4 +1,15 @@
 import { PenBoxIcon, TrashIcon } from "lucide-react";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "./ui/alert-dialog";
 
 interface ActionButtonProps extends React.HTMLAttributes<HTMLDivElement> {
   onEditClick: () => void;
@@ -18,12 +29,31 @@ export default function ActionButton({
       >
         <PenBoxIcon className="h-4 w-4 xs:h-3 xs:w-3" />
       </div>
-      <div
-        className="flex h-6 w-6 items-center justify-center rounded-md bg-red-300/30 hover:bg-red-300/50"
-        onClick={() => onDeleteClick()}
-      >
-        <TrashIcon className="h-4 w-4 xs:h-3 xs:w-3" />
-      </div>
+      <AlertDialog>
+        <AlertDialogTrigger>
+          <div className="flex h-6 w-6 items-center justify-center rounded-md bg-red-300/30 hover:bg-red-300/50">
+            <TrashIcon className="h-4 w-4 xs:h-3 xs:w-3" />
+          </div>
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>Are you absolutely sure?</AlertDialogTitle>
+            <AlertDialogDescription>
+              This action cannot be undone. This will permanently delete your
+              account and remove your data from our servers.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Cancel</AlertDialogCancel>
+            <AlertDialogAction
+              className="bg-red-300/50 hover:bg-red-300/30"
+              onClick={() => onDeleteClick()}
+            >
+              Continue
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     </div>
   );
 }
